pyproject.toml
[project]
name = "lilith-gmail"
version = "0.1.0"
description = "Gmail sync daemon + Lilith agent tools for semantic email search"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.3",
    "rich>=13.9.4",
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.4",
    "fastapi>=0.128.1",
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "mcp>=1.0.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/sync", "src/transform", "src/daemon", "src/mcp_server"]

[project.scripts]
lilith-gmail = "main:main"
mcp = "mcp_server.__main__:main"
